Rename pptx.py to pptx_addon.py to prevent a naming collision
